requirements.txt
fastapi
uvicorn
deta
python-dotenv
python-jose[cryptography]
passlib
bcrypt
oauth2
aiofiles
python-multipart
pydantic